refactor: move functions out of embrace ext
