+++ embrace-android-otel/src/main/kotlin/io/embrace/android/embracesdk/internal/otel/spans/SpanExt.kt
@@ -1,0 +1,32 @@
+package io.embrace.android.embracesdk.internal.otel.spans
+
+import io.embrace.android.embracesdk.internal.clock.millisToNanos
+import io.embrace.android.embracesdk.internal.otel.attrs.EmbraceAttribute
+import io.embrace.android.embracesdk.internal.otel.schema.AppTerminationCause
+import io.embrace.android.embracesdk.internal.otel.schema.EmbType
+import io.embrace.android.embracesdk.internal.otel.schema.ErrorCodeAttribute
+import io.embrace.android.embracesdk.internal.otel.sdk.hasEmbraceAttribute
+import io.embrace.android.embracesdk.internal.otel.sdk.setEmbraceAttribute
+import io.embrace.android.embracesdk.internal.payload.Attribute
+import io.embrace.android.embracesdk.internal.payload.Span
+import io.opentelemetry.api.trace.SpanId
+
+fun Span.hasEmbraceAttribute(embraceAttribute: EmbraceAttribute): Boolean {
+    return embraceAttribute.value == attributes?.singleOrNull { it.key == embraceAttribute.key.name }?.data
+}
+
+fun Span.toFailedSpan(endTimeMs: Long): Span {
+    val newAttributes = mutableMapOf<String, String>().apply {
+        setEmbraceAttribute(ErrorCodeAttribute.Failure)
+        if (hasEmbraceAttribute(EmbType.Ux.Session)) {
+            setEmbraceAttribute(AppTerminationCause.Crash)
+        }
+    }
+
+    return copy(
+        endTimeNanos = endTimeMs.millisToNanos(),
+        parentSpanId = parentSpanId ?: SpanId.getInvalid(),
+        status = Span.Status.ERROR,
+        attributes = newAttributes.map { Attribute(it.key, it.value) }.plus(attributes ?: emptyList())
+    )
+}
